requirements.txt
certifi==2019.9.11
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.0
PyMySQL==0.9.3
python-slugify==3.0.0
SQLAlchemy==1.3.1
text-unidecode==1.2
Werkzeug==0.16.0
WTForms==2.2.1
